third_party/WebKit/LayoutTests/device_orientation/orientation/optional-event-properties.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description('Tests the optional properties of DeviceOrientationEvent. Each property should be null if not set, or set to null or undefined.');

var event;

evalAndLog("event = document.createEvent('DeviceOrientationEvent')");
shouldBeTrue("event.alpha == null");
shouldBeTrue("event.beta == null");
shouldBeTrue("event.gamma == null");
shouldBeFalse("event.absolute");

evalAndLog("event.initDeviceOrientationEvent('', false, false, 0, 1, 2, false)");
shouldBeTrue("event.alpha == 0");
shouldBeTrue("event.beta == 1");
shouldBeTrue("event.gamma == 2");
shouldBeFalse("event.absolute");

evalAndLog("event.initDeviceOrientationEvent('', false, false, 0, 1, 2, true)");
shouldBeTrue("event.alpha == 0");
shouldBeTrue("event.beta == 1");
shouldBeTrue("event.gamma == 2");
shouldBeTrue("event.absolute");

evalAndLog("event.initDeviceOrientationEvent()");
shouldBeTrue("event.alpha == null");
shouldBeTrue("event.beta == null");
shouldBeTrue("event.gamma == null");
shouldBeFalse("event.absolute");

evalAndLog("event.initDeviceOrientationEvent('', false, false, [], [], [], [])");
shouldBeTrue("event.alpha == 0");
shouldBeTrue("event.beta == 0");
shouldBeTrue("event.gamma == 0");
shouldBeTrue("event.absolute");

evalAndLog("event.initDeviceOrientationEvent('', false, false, undefined, undefined, undefined, undefined)");
shouldBeTrue("event.alpha == null");
shouldBeTrue("event.beta == null");
shouldBeTrue("event.gamma == null");
shouldBeFalse("event.absolute");

evalAndLog("event.initDeviceOrientationEvent('', false, false, '', '', '', '')");
shouldBeTrue("event.alpha == 0");
shouldBeTrue("event.beta == 0");
shouldBeTrue("event.gamma == 0");
shouldBeFalse("event.absolute");

evalAndLog("event.initDeviceOrientationEvent('', false, false, null, null, null, null)");
shouldBeTrue("event.alpha == null");
shouldBeTrue("event.beta == null");
shouldBeTrue("event.gamma == null");
shouldBeFalse("event.absolute");
</script>
</body>
</html>
